Add stress test
